refactor: reorganize folder structure for better module separation
- Separated and categorized files based on feature, functionality, and purpose
- Designated specific folders for components, pages, services, pipes, models, utilities and etc.)
- Improved overall project structure to support scalability and maintainability
