+++ src/app/pages/user-details/user-details.component.scss
