+++ src/app/pages/users-list/users-list.component.ts
@@ -1,5 +1,5 @@
 import {Component} from '@angular/core';
-import {UserService} from 'src/app/services/user.service';
+import {UserService} from 'src/app/services/user/user.service';
 import { CommonModule } from '@angular/common';
 
 @Component({
