+++ src/model/index.ts
@@ -1,0 +1,2 @@
+export * from  './new-user-data';
+export * from './user-data';
